01-intro-to-mcp-agents/code/requirements.txt
mcp
httpx
python-dotenv
pydantic
